sync optimized and some functions refactored

+++ src/lib/RestAPI.js
@@ -57,8 +57,8 @@ class RestAPI {
    */
   async _fetch(opts, fetchRequestId) {
     let url = opts.url
-    if(!_.includes(opts.url, CONFIG.API.SEND_SMS_LINK) && !_.includes(opts.url, CONFIG.API.SEND_EMAIL_LINK)){
-    url = this.API_BASE_URL + url
+    if (!_.includes(opts.url, CONFIG.API.SEND_SMS_LINK) && !_.includes(opts.url, CONFIG.API.SEND_EMAIL_LINK)) {
+      url = this.API_BASE_URL + url
     }
     if (this._sessionToken) {
       opts.headers['Cookie'] = this._sessionToken
@@ -188,8 +188,6 @@ class RestAPI {
   }
 
   async uploadZipFile(path, fileName) {
-    // const jid = this._sessionToken.split(';')[1].split(',')[1].trim()
-    // console.log('jid',jid)
     var PATH = (!path) ? RNFS.DocumentDirectoryPath + '/' + CONFIG.APP_FOLDER : path
     var filePath = (!path) ? PATH + '/sync.zip' : PATH
     let responseBody = "Fail"
@@ -200,27 +198,15 @@ class RestAPI {
         { name: 'file', filename: (!fileName) ? 'sync.zip' : fileName, data: RNFetchBlob.wrap(filePath) },
       ]).uploadProgress((written, total) => {
       }).then(async (resp) => {
-        if (!path) {
-          responseBody = resp.text()
-          const message = responseBody.split(",")[0]
-          const syncCount = responseBody.split(",")[1]
-          if (message == 'success') {
-            //do something
-            const currenDate = moment(new Date()).format('YYYY-MM-DD HH:mm:ss')
-            await keyValueDBService.validateAndSaveData(LAST_SYNC_WITH_SERVER, currenDate)
-            await keyValueDBService.deleteValueFromStore(PENDING_SYNC_TRANSACTION_IDS);
-            // let transactionIdToBeSynced = await keyValueDBService.getValueFromStore(PENDING_SYNC_TRANSACTION_IDS);
-          }
-          else {
-            throw new Error(responseBody)
-          }
-        } else {
-          responseBody = resp.text()
-          console.log('responseBody', responseBody)
-          const message = responseBody.split(",")[0]
-          if (message != 'success') {
-            throw new Error(responseBody)
-          }
+        responseBody = resp.text()
+        const message = responseBody.split(",")[0]
+        if (!path && message == 'success') {
+          const currenDate = moment(new Date()).format('YYYY-MM-DD HH:mm:ss')
+          await keyValueDBService.validateAndSaveData(LAST_SYNC_WITH_SERVER, currenDate)
+          await keyValueDBService.deleteValueFromStore(PENDING_SYNC_TRANSACTION_IDS);
+        }
+        else if (message != 'success') {
+          throw new Error(responseBody)
         }
       }).catch(err => {
         throw {
